+++ packages/werkbank/lib/src/_internal/src/widgets/src/use_case_widget_display.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:werkbank/src/_internal/src/widgets/widgets.dart';
+import 'package:werkbank/src/use_case/use_case.dart';
 
 /// Displays the use case defined controller in
 /// [UseCaseControllerProvider].
@@ -14,10 +15,13 @@ class UseCaseWidgetDisplay extends StatelessWidget {
     return ListenableBuilder(
       listenable: controller,
       builder: (context, _) {
-        return controller.useCaseWidget(
-          errorBuilder: (context, error, _) {
-            return ErrorWidget(error);
-          },
+        return KeyedSubtree(
+          key: UseCase.key,
+          child: controller.useCaseWidget(
+            errorBuilder: (context, error, _) {
+              return ErrorWidget(error);
+            },
+          ),
         );
       },
     );