+++ packages/werkbank/lib/src/_internal/src/widgets/widgets.dart
@@ -22,6 +22,7 @@ export 'src/panels/navigation_panel/sliver_pinned_header.dart';
 export 'src/providers/descriptor_provider.dart';
 export 'src/providers/is_warm_start_provider.dart';
 export 'src/providers/json_store_provider.dart';
+export 'src/providers/theme_brightness_provider.dart';
 export 'src/providers/use_case_composition/use_case_composition_by_controller_provider.dart';
 export 'src/providers/use_case_composition/use_case_composition_provider.dart';
 export 'src/providers/use_case_controller_provider/current_use_case_controller_by_manager_provider.dart';