+++ packages/werkbank/test/display_test.dart
@@ -19,6 +19,8 @@ void main() {
             useCase: useCases,
           ),
         );
+
+        expect(find.byKey(UseCase.key), findsOneWidget);
       });
     }
   });